fix(ir): negative list index corrupted the heap in the IR interpreter
The IR interpreter's ListIndexStore opcode passed the index straight to
QoreListNode::setEntry(size_t).  A negative index arrived there as SIZE_MAX,
so "index + 1" wrapped to 0, resize() was a no-op, and the store wrote 8 bytes
before the start of the list's entry array, silently corrupting heap metadata;
the process then aborted at an unrelated later free.  The JIT and AOT runtime
helpers already resolved negative indices against the list size (or raised
NEGATIVE-LIST-INDEX); only the IR tier was missing the check, so any
"l[-1] = v" with a literal index -- the shape that lowers to ListIndexStore --
was enough to corrupt the heap.

Exceptions raised inside the same opcode also bypassed the enclosing catch
block in the IR and JIT tiers: the lowering never set the instruction's
exception target, and the interpreter's error paths returned out of execute()
instead of branching to the handler.  The store's container-slot cleanup now
runs on the error path as well, as the equivalent hash store already does,
so the borrowed container reference is not released twice.

QoreListNode::setEntry() now rejects an index it cannot resize to hold rather
than writing outside the array.

Also widen source line numbers from 16 to 32 bits.  QoreProgramLineLocation
stored start/end lines in an int16_t, so any file over 32767 lines wrapped to
negative line numbers in stack traces, exception locations, and assertions;
the AOT wire format truncated them to u16 as well, which the new
QORE_AOT_FEAT_WIDE_LOC_LINES feature bit fixes.

+++ include/qore/intern/QoreIR.h
@@ -836,7 +836,10 @@ public:
     virtual ~QoreIRInstruction() = default;
 
     QoreIROpcode opcode;
-    int16_t cached_start_line;  // -1 = no loc, >=0 = loc->start_line (fills padding after opcode)
+    // -1 = no loc, >=0 = loc->start_line (fills the padding after the opcode, so 32 bits are free here); must be
+    // 32-bit to match QoreProgramLineLocation::start_line, or lines > 32767 wrap to negative values and the
+    // debug line-change gate below reports the wrong line
+    int32_t cached_start_line;
     const QoreProgramLocation* loc = nullptr;
     QoreIRValue result{};
     std::vector<QoreIRValue> operands;
